Add github.actions.workflows.setup_version package
Check that `*-version' in the `setup-*' actions are strings.

When using the various `setup-*' actions due YAML can truncate possible
trailing 0-s ending up using Go 1.2 (instead of Go 1.20) or Python 3.1
(instead of Python 3.10).

In some cases this is not easy to debug.

Add conftest rules to spot such cases.