Switch to use `some ... in' again
Switch to use `some ... in' again, this time hopefully correctly.

In commit 35314ce23f7676b95fae18f73f9577aff2f0da03 there were the
problem that by using only a single argument with `in' operator, we
actually unified with the actual object, not its key.

- Further checks were wrong because they assumed the key, not the object
- This made reporting when a violation happened more verbose (entire
  object was printed instead of just the interesting key)

Use the `in' operator with two arguments in order to get the key and
preserve the behaviour that we had.

Closes issue #29.