Add script to generate package(s) documentation
Accept one or more paths as argument and pass all its argument to `opa
inspect -a', massage via jq and generate a markdown documented that
could be used as README.md in order to document all packages contained
in such paths.

Intended to be used in order to automatically (re)generate such
README.md-s so that we have all packages automatically documented.