Run docs target to double-check for updated documentation
Run `make docs' and double-check that it is updated by inspecting the
output of `git status --porcelain'.

If the documentation is not updated some README.md will be modified or
created and the awk pipe will exit with a non-0 status.

As a side-effect this also triggers if for any reason the repository is
"dirty".