Create temporary bundles without test files
Switch to create temporary bundles that do not have Rego test files.

In conftest test files we are using `parse_config()` and
`parse_config_file()` that are not present in OPA and in recent OPA
version doing that ends up with possible unsafe variable uses.

Add a big XXX comment regarding that.

(Unfortunately using stdout for bundle is not possible right now so we
need to store all of that in a temporary file.)